// typescript-projects/utility-backups/src20250516BeforeErrorHandlingChange/interface/types.ts
export type TEbillsBetPaymentData = {
  provider: ProviderId;
  customerId: string;
  merchant: string;
  amount: number; // `amount` in tx doc root
};